Prettier code
Going to start looking at TypeScript.
