+++ src/client/pages/Landing.jsx
@@ -1,21 +1,18 @@
 import React from "react";
-import NavBar from "../components/NavBar";
-import Links from "../components/Links";
 import "../../shared/css/App.css";
 import "../../shared/css/FadeInSection.css";
-
-
+import Background from "../components/Links";
+import NavBar from "../components/Navbar";
 import Hero from "../sections/Hero";
 import Projects from "../sections/Projects";
-import Contact from "../sections/Contact";
 import Skills from "../sections/Skills";
+import Contact from "../sections/Contact";
 import Footer from "../sections/Footer";
 
 export default function Landing () {
-
     return (
         <div className="App">
-            <Links/>
+            <Background/>
             <NavBar/>
                 <Hero/>
                 <Projects/>
@@ -23,4 +20,5 @@ export default function Landing () {
                 <Contact/>
             <Footer/>
         </div>
-)}
+    );
+}
