+++ src/client/sections/Hero.jsx
@@ -1,16 +1,14 @@
-import SendRoundedIcon from "@mui/icons-material/SendRounded";
 import React from "react";
+import { Link } from "react-scroll";
 import FadeInSection from "../lib/FadeInSection";
+import SendRoundedIcon from "@mui/icons-material/SendRounded";
 import githubLogo from "../../shared/img/GitHub_Logo_White.png";
 import githubMark from "../../shared/img/GitHub-Mark-Light-64px.png";
-import {Link} from "react-scroll";
-
 
 export default function Hero () {
     return (
         <div className="hero">
             <div className="text-zone">
-
                 <div>
                     <FadeInSection type={"fade-in-3"} >
                         <h3 className="">Hi, my name is</h3>
@@ -30,8 +28,8 @@ export default function Hero () {
 
                         <a href="https://github.com/MSkauen">
                             <button className="btn">
-                                <img className="logo" src={githubLogo}/>
-                                <img className="mark" src={githubMark}/>
+                                <img className="logo" src={githubLogo} alt="Github"/>
+                                <img className="mark" src={githubMark} alt="GitHub Logo"/>
                             </button>
                         </a>
 
@@ -51,6 +49,5 @@ export default function Hero () {
                 </div>
             </div>
         </div>
-
-    )
+    );
 }
