+++ src/client/sections/Projects.jsx
@@ -1,14 +1,12 @@
 import React from "react";
 
-
 export default function Projects () {
     return (
         <div className="projects">
             <div className="section">
                 <div className="project-innerRef"/>
-                <h1>
-                    Projects
-                </h1>
+                <h1>Projects</h1>
+
                 <div className="project-grid">
                     <div className="card"/>
                     <div className="card"/>
@@ -16,5 +14,5 @@ export default function Projects () {
                 </div>
             </div>
         </div>
-    )
+    );
 }
