+++ src/client/sections/Space.jsx
@@ -2,8 +2,6 @@ import React from "react";
 
 export default function Space () {
     return (
-        <div className="space">
-
-        </div>
-    )
+        <div className="space"/>
+    );
 }
