+++ src/shared/html/index.html
@@ -1,13 +1,13 @@
 <!DOCTYPE html>
 <html lang="en">
-<head>
-    <meta charset="UTF-8" />
-    <meta name="viewport" content="width=device-width, initial-scale=1, minimum-scale=1 viewport-fit=cover" />
-    <title>Mathias Skauen Harestad</title>
-    <link rel="icon" type="image/x-icon" href="../favicon.ico">
-</head>
-<body>
-<div id="root"></div>
-</body>
-<script src="../../index.jsx" type="module"></script>
+    <head>
+        <meta charset="UTF-8"/>
+        <meta name="viewport" content="width=device-width, initial-scale=1, minimum-scale=1 viewport-fit=cover"/>
+        <title>Mathias Skauen Harestad</title>
+        <link rel="icon" type="image/x-icon" href="../favicon.ico">
+    </head>
+    <body>
+        <div id="root"></div>
+    </body>
+    <script src="../../index.jsx" type="module"></script>
 </html>
